Implemented importing of raw_analytics only
It is now possible to work purely based on the raw_analytics files. We had this ability way back, but then it seemed likely we would use either the datashop file or the raw_analytics and datashop in combination, so we ended up removing it. Now I have put it back in.