Added boundaries to test.
We now consider participants to have had poor skills if they started out getting worse than 50% of questions right on the first try, and to have ended up with good skills if they were able to answer more than 80% right in the end. We consider aversion of the module to be significantly better if the probability of moving a poor performer into the group of high performers is 20% higher than for the other version.